Fixed 64-bit compiler warnings
git-svn-id: https://svn.code.sf.net/p/apprenticevideo/code/trunk@232 28900df4-563f-46a4-9db8-70c0268b8f86

+++ apprenticevideo/yaeReaderFFMPEG.cpp
@@ -1194,7 +1194,7 @@ namespace yae
         for (unsigned char i = 0; i < numSamplePlanes_; i++)
         {
           pict.data[i] = sampleBuffer->samples(i);
-          pict.linesize[i] = sampleBuffer->rowBytes(i);
+          pict.linesize[i] = (int)sampleBuffer->rowBytes(i);
         }
         
         sws_scale(imgConvertCtx_,
